♻️REFACTOR: LottoResult 패키지 이동
