+++ src/test/java/lotto/domain/lotto/WinLottoTest.java
@@ -3,6 +3,7 @@ package lotto.domain.lotto;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import lotto.domain.lottoresult.LottoResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
